+++ .gitignore
@@ -3,3 +3,4 @@ dist
 .vscode/
 .DS_Store
 *scss\.d\.ts
+coverage
